Adjusted some things for better compatibility with CRT televisions
Adjusted the delay of the run-in clock to more closely match the specs that CC decoders expect.  It is supposed to start at 10.5us from the falling edge of the HSYNC pulse.

Decoders in older TVs also seem to expect a constant CC signal even when no data is being transmitted.  The routine has been changed so that there is a constant signal being sent with zeros as the data bytes when it is idling.